fix : 패치 삭제 오류 수정

+++ backend/src/main/java/com/example/backend/domain/fetch/service/FetchServiceImpl.java
@@ -116,6 +116,9 @@ public class FetchServiceImpl implements FetchService {
         } else {
             log.info("fetch get{}", fetch.get());
             Optional<MemberFetch> memberFetch = memberFetchRepository.findByFetchAndMember(fetch.get(), member);
+            if(memberFetch.isEmpty()) {
+                throw new BusinessException(ErrorCode.FETCH_NOT_FOUND);
+            }
             memberFetchRepository.delete(memberFetch.get());
         }
